// daily-shop/src/app/model/product.ts
export class Product{
    id: number;
    title: string;
    description: string;
    img: string;
    size: string;
    price: number;
    type: string;
    avilability: string;
    category: string;
}


// avilability: "In stock"
// category: "Polo T-Shirt"
// description: 
// id: 1
// img: "https://www.free-css.com/assets/files/free-css-..."
// price: 34.99
// size: "S"
// title:"T-Shirt"
// type: "men"